+++ src/app/infuse/infuse.component.ts
@@ -1,6 +1,6 @@
-import { extend, copy, IComponentOptions, IController, IQService } from 'angular';
-import * as _ from 'underscore';
-import { flatMap } from '../util';
+import { IComponentOptions, IController, IQService } from 'angular';
+import copy from 'fast-copy';
+import * as _ from 'lodash';
 import { getDefinitions } from '../destiny1/d1-definitions.service';
 import template from './infuse.html';
 import './infuse.scss';
@@ -33,7 +33,7 @@ function InfuseCtrl(
 
   const vm = this;
 
-  extend(vm, {
+  Object.assign(vm, {
     items: {},
     getAllItems: true,
     showLockedItems: false,
@@ -101,7 +101,7 @@ function InfuseCtrl(
       }
 
       if (vm.query.infusable) {
-        let targetItems = flatMap(stores, (store) => {
+        let targetItems = _.flatMap(stores, (store) => {
           return store.items.filter((item) => {
             return (!item.locked || vm.showLockedItems) && vm.isInfusable(vm.query, item);
           });
@@ -113,7 +113,7 @@ function InfuseCtrl(
       }
 
       if (vm.query.infusionFuel) {
-        let sourceItems = flatMap(stores, (store) => {
+        let sourceItems = _.flatMap(stores, (store) => {
           return store.items.filter((item) => {
             return vm.isInfusable(item, vm.query);
           });