+++ src/app/search/search-filters.ts
@@ -1,5 +1,5 @@
-import * as _ from 'underscore';
-import { flatMap } from '../util';
+import * as _ from 'lodash';
+
 import { compareBy, chainComparator, reverseComparator } from '../comparators';
 import { DimItem, D1Item, D2Item } from '../inventory/item-types';
 import { StoreServiceType, DimStore } from '../inventory/store-types';
@@ -63,7 +63,7 @@ interface SearchConfig {
  */
 export function buildSearchConfig(destinyVersion: 1 | 2): SearchConfig {
   const categories = destinyVersion === 1 ? D1Categories : D2Categories;
-  const itemTypes = flatMap(Object.values(categories), (l: string[]) =>
+  const itemTypes = _.flatMap(Object.values(categories), (l: string[]) =>
     l.map((v) => v.toLowerCase())
   );
 
@@ -274,6 +274,7 @@ export function buildSearchConfig(destinyVersion: 1 | 2): SearchConfig {
 
   // free form notes on items
   keywords.push('notes:');
+  keywords.push('perk:');
 
   // Build an inverse mapping of keyword to function name
   const keywordToFilter: { [key: string]: string } = {};
@@ -479,7 +480,7 @@ export function searchFilters(
       // with the previous one in a hacked-up "or" node that we'll handle specially.
       let or = false;
 
-      function addPredicate(predicate: string, filter: string, invert = false) {
+      function addPredicate(predicate: string, filter: string, invert: boolean = false) {
         const filterDef: Filter = { predicate, value: filter, invert };
         if (or && filters.length) {
           const lastFilter = filters.pop();
@@ -495,8 +496,11 @@ export function searchFilters(
         or = false;
       }
 
-      for (let term of searchTerms) {
-        term = term.replace(/['"]/g, '');
+      for (let search of searchTerms) {
+        search = search.replace(/['"]/g, '');
+
+        const invert = search.startsWith('-');
+        const term = search.replace(/^-/, '');
 
         if (term === 'or') {
           or = true;
@@ -504,41 +508,44 @@ export function searchFilters(
           const filter = term.replace('is:', '');
           const predicate = searchConfig.keywordToFilter[filter];
           if (predicate) {
-            addPredicate(predicate, filter);
+            addPredicate(predicate, filter, invert);
           }
         } else if (term.startsWith('not:')) {
           const filter = term.replace('not:', '');
           const predicate = searchConfig.keywordToFilter[filter];
           if (predicate) {
-            addPredicate(predicate, filter, true);
+            addPredicate(predicate, filter, !invert);
           }
         } else if (term.startsWith('tag:')) {
           const filter = term.replace('tag:', '');
-          addPredicate('itemtags', filter);
+          addPredicate('itemtags', filter, invert);
         } else if (term.startsWith('notes:')) {
           const filter = term.replace('notes:', '');
-          addPredicate('notes', filter);
+          addPredicate('notes', filter, invert);
+        } else if (term.startsWith('perk:')) {
+          const filter = term.replace('perk:', '');
+          addPredicate('perk', filter, invert);
         } else if (term.startsWith('light:') || term.startsWith('power:')) {
           const filter = term.replace('light:', '').replace('power:', '');
-          addPredicate('light', filter);
+          addPredicate('light', filter, invert);
         } else if (term.startsWith('masterwork:')) {
           const filter = term.replace('masterwork:', '');
-          addPredicate('masterworkValue', filter);
+          addPredicate('masterworkValue', filter, invert);
         } else if (term.startsWith('stack:')) {
           const filter = term.replace('stack:', '');
-          addPredicate('stack', filter);
+          addPredicate('stack', filter, invert);
         } else if (term.startsWith('level:')) {
           const filter = term.replace('level:', '');
-          addPredicate('level', filter);
+          addPredicate('level', filter, invert);
         } else if (term.startsWith('quality:') || term.startsWith('percentage:')) {
           const filter = term.replace('quality:', '').replace('percentage:', '');
-          addPredicate('quality', filter);
+          addPredicate('quality', filter, invert);
         } else if (term.startsWith('rating:')) {
           const filter = term.replace('rating:', '');
-          addPredicate('rating', filter);
+          addPredicate('rating', filter, invert);
         } else if (term.startsWith('ratingcount:')) {
           const filter = term.replace('ratingcount:', '');
-          addPredicate('ratingcount', filter);
+          addPredicate('ratingcount', filter, invert);
         } else if (term.startsWith('stat:')) {
           // Avoid console.error by checking if all parameters are typed
           const pieces = term.split(':');
@@ -547,7 +554,7 @@ export function searchFilters(
             addPredicate(filter, pieces[2]);
           }
         } else if (!/^\s*$/.test(term)) {
-          addPredicate('keyword', term.replace(/^-/, ''), term.startsWith('-'));
+          addPredicate('keyword', term, invert);
         }
       }
 
@@ -638,7 +645,7 @@ export function searchFilters(
         return item.talentGrid && item.talentGrid.hasAscendNode && !item.talentGrid.ascended;
       },
       reforgeable(item: DimItem) {
-        return item.talentGrid && _.any(item.talentGrid.nodes, { hash: 617082448 });
+        return item.talentGrid && item.talentGrid.nodes.some((n) => n.hash === 617082448);
       },
       ornament(item: D1Item, predicate: string) {
         const complete = item.talentGrid && item.talentGrid.nodes.some((n) => n.ornament);
@@ -811,7 +818,7 @@ export function searchFilters(
       stattype(item: DimItem, predicate: string) {
         return (
           item.stats &&
-          _.any(item.stats, (s) =>
+          item.stats.some((s) =>
             Boolean(s.name.toLowerCase() === predicate && s.value && s.value > 0)
           )
         );
@@ -844,6 +851,11 @@ export function searchFilters(
           // Search for typeName (itemTypeDisplayName of modifications)
           item.typeName.toLowerCase().includes(predicate) ||
           // Search perks as well
+          this.perk(item, predicate)
+        );
+      },
+      perk(item: DimItem, predicate: string) {
+        return (
           (item.talentGrid &&
             item.talentGrid.nodes.some((node) => {
               // Fixed #798 by searching on the description too.
@@ -1111,7 +1123,7 @@ export function searchFilters(
       hasShader(item: D2Item) {
         return (
           item.sockets &&
-          _.any(item.sockets.sockets, (socket) => {
+          item.sockets.sockets.some((socket) => {
             return (
               (socket.plug || false) &&
               socket.plug.plugItem.plug.plugCategoryHash === 2973005342 &&