+++ src/app/settings/SettingsPage.tsx
@@ -11,9 +11,8 @@ import exampleWeaponImage from 'app/images/example-weapon.jpg';
 // tslint:disable-next-line:no-implicit-dependencies
 import exampleArmorImage from 'app/images/example-armor.jpg';
 import './settings.scss';
-import { $rootScope } from 'ngimport';
 import { DimItem } from '../inventory/item-types';
-import * as _ from 'underscore';
+import * as _ from 'lodash';
 import { reviewPlatformOptions } from '../destinyTrackerApi/platformOptionsFetcher';
 import { getReviewModes } from '../destinyTrackerApi/reviewModesFetcher';
 import { downloadCsvFiles } from '../inventory/dimCsvService.factory';
@@ -198,13 +197,16 @@ class SettingsPage extends React.Component<Props, State> {
     }
 
     const itemSortCustom = _.sortBy(
-      _.map(itemSortProperties, (displayName, id) => {
-        return {
-          id,
-          displayName,
-          enabled: sortOrder.includes(id)
-        };
-      }),
+      _.map(
+        itemSortProperties,
+        (displayName, id): SortProperty => {
+          return {
+            id,
+            displayName,
+            enabled: sortOrder.includes(id)
+          };
+        }
+      ),
       (o) => {
         const index = sortOrder.indexOf(o.id);
         return index >= 0 ? index : 999;
@@ -547,9 +549,6 @@ class SettingsPage extends React.Component<Props, State> {
 
     localStorage.setItem('dimLanguage', language);
     changeLanguage(language, () => {
-      $rootScope.$applyAsync(() => {
-        $rootScope.$broadcast('i18nextLanguageChange');
-      });
       this.setState({});
     });
   };