+++ src/app/shell/refresh.tsx
@@ -1,11 +1,20 @@
 import * as React from 'react';
 import { hotkeys } from '../ngimport-more';
-import { $rootScope } from 'ngimport';
 import { t } from 'i18next';
 import classNames from 'classnames';
 import { loadingTrackerStream } from './dimLoadingTracker.factory';
 import { Subscription } from 'rxjs/Subscription';
 import { AppIcon, refreshIcon } from './icons';
+import { Subject } from 'rxjs/Subject';
+
+export const refresh$ = new Subject();
+
+export function refresh() {
+  // Individual pages should listen to this event and decide what to refresh,
+  // and their services should decide how to cache/dedup refreshes.
+  // This event should *NOT* be listened to by services!
+  refresh$.next();
+}
 
 export default class Refresh extends React.Component<{}, { active: boolean }> {
   private subscription: Subscription;
@@ -19,9 +28,7 @@ export default class Refresh extends React.Component<{}, { active: boolean }> {
     hotkeys.add({
       combo: ['r'],
       description: t('Hotkey.RefreshInventory'),
-      callback: () => {
-        this.refresh();
-      }
+      callback: refresh
     });
 
     this.subscription = loadingTrackerStream.subscribe((active) => {
@@ -38,16 +45,9 @@ export default class Refresh extends React.Component<{}, { active: boolean }> {
     const { active } = this.state;
 
     return (
-      <span className="link" onClick={this.refresh} title={t('Header.Refresh')}>
+      <span className="link" onClick={refresh} title={t('Header.Refresh')}>
         <AppIcon icon={refreshIcon} className={classNames({ 'fa-spin': active })} />
       </span>
     );
   }
-
-  private refresh = () => {
-    // Individual pages should listen to this event and decide what to refresh,
-    // and their services should decide how to cache/dedup refreshes.
-    // This event should *NOT* be listened to by services!
-    $rootScope.$broadcast('dim-refresh');
-  };
 }