+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next
 
+* Added negative search. Prefix any search term with `-` and it will match the opposite.
+* Added `perk:"* **"` seach filter to match any keywords against perks on an item
+* Lock and unlock items matching your current search from the same menu you use for tagging them.
+
 # 4.73.0 (2018-10-07)
 
 * Added `is:heroic` search filter for armor with heroic resistance.